[Certora Audit] I-02. Some comments say `keccak` instead of `keccak256` (#886)
This pull request includes updates to comments in the Solidity smart
contracts to clarify the usage of the `keccak256` function. The changes
ensure that the comments accurately reflect the encoding process used in
the contracts.

Changes to comments for clarity:

*
[`contracts/common/SecuredTokenTransfer.sol`](diffhunk://#diff-7a34930a339acfe3b45e163bee3e08df2132c01826e6e03771827a4181c6f567L19-R19):
Updated the comment to specify that `0xa9059cbb` is the `bytes4`
encoding of `keccak256("transfer(address,uint256)")`.
*
[`contracts/proxies/SafeProxy.sol`](diffhunk://#diff-754f853ea53666aa85b2d27bbcc0623b7cfd83449e0b949eed39dde8f01ba1cdL41-R41):
Updated the comment to specify that `0xa619486e` is the `bytes4`
encoding of `keccak256("masterCopy()")`.

Co-authored-by: Mikhail <16622558+mmv08@users.noreply.github.com>