[Certora Audit] G-03. `ERC165Handler.setSupportedInterface()`: Logic and storage access optimization (#890)
This pull request includes an important update to the `ERC165Handler`
contract to improve the handling of interface support and streamline the
code. The most significant changes are as follows:

### The following optimizes the logic and minimizes storage accesses:

*
[`contracts/handler/extensible/ERC165Handler.sol`](diffhunk://#diff-aa0838f20fd3f37b00dc661645b4641500e68762b9b624addb99465fcc65a3e0L38-R48):
Refactored the logic for checking and updating the support status of
interfaces by introducing a local `mapping` reference. This change
simplifies the code and ensures that the interface support status is
updated correctly.