[Certora Audit] G-06. `ExtensibleFallbackHandler._supportsInterface()`: save gas via short-circuit evaluation (#893)
This pull request includes a change to the `ExtensibleFallbackHandler`
contract in the `contracts/handler/ExtensibleFallbackHandler.sol` file.
The change modifies the `_supportsInterface` function to reorder the
interface checks for `ERC721TokenReceiver`, `ERC1155TokenReceiver`, and
`IFallbackHandler`. This helps in taking advantage of the short-circuit
evaluation.

Changes to interface support order:

*
[`contracts/handler/ExtensibleFallbackHandler.sol`](diffhunk://#diff-aa345618c4d3f173b09e211d0bd0eec0747177aab345bf8b9f5bbc874a765fe3R21-R26):
Reordered the interface checks in the `_supportsInterface` function to
place `ERC721TokenReceiver` and `ERC1155TokenReceiver` before
`IFallbackHandler`.