Fix links (#931)
The original master file has been changed, and the line numbers
referenced by the original URL are no longer accurate. Therefore, a
fixed permanent link is used to replace it. The purpose of adopting a
permanent link is to prevent the line numbers originally pointed to by
the URL from becoming inaccurate again when further changes are made to
the master branch.